requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.1.0
executing==0.8.3
fonttools==4.33.1
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nltk==3.7
numpy==1.22.3
oauth2client==4.1.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2==2.9.3
pure-eval==0.2.2
pyagrum==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
PyDrive==1.3.1
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
static3==0.7.0
tqdm==4.64.0
traitlets==5.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
